billing_service/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==4.0.3
asyncpg==0.30.0
attrs==24.2.0
babel==2.16.0
beanie==1.27.0
black==24.10.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
choicesenum==0.7.0
click==8.1.7
cryptography==43.0.3
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.2
fastapi-cli==0.0.5
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
idna==3.10
Jinja2==3.1.4
lazy-model==0.2.0
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
motor==3.6.0
multidict==6.1.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
phonenumbers==8.13.47
platformdirs==4.3.6
propcache==0.2.0
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.0
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.9.0
pymongo==4.9.2
python-dotenv==1.0.1
python-multipart==0.0.12
PyYAML==6.0.2
requests==2.32.3
rich==13.9.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
SQLAlchemy-Utils==0.41.2
starlette==0.40.0
stripe==11.2.0
toml==0.10.2
tomli==2.0.2
typer==0.12.5
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
uvloop==0.21.0
watchfiles==0.24.0
websockets==13.1
yarl==1.17.1
